pyproject.toml
[project]
name = "mcq-engine"
version = "0.1.0"
description = "Monte Carlo Quantitative Trading Engine for NQ/MNQ futures"
requires-python = ">=3.11"
dependencies = [
    "polars>=0.20",
    "pandas>=2.1",
    "pyarrow>=14",
    "numpy>=1.26",
    "scipy>=1.12",
    "plotly>=5.18",
    "matplotlib>=3.8",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.pytest.ini_options]
testpaths = ["tests"]
